encoder:mudança do tipo da variável pulse

+++ firmware/lib/encoder/encoder.h
@@ -32,7 +32,7 @@ typedef struct
 {
     union velocity speed;
 
-    float pulse;
+    int pulse;
 
     encoder_params_t gear_box;
